Improve door generation
